ZETA-4435: Move over scaffolds to it's own standalone function.
