+++ src/extension.ts
@@ -11,17 +11,7 @@ import { onVSCodeClose, tryToAuth, updatePrivateDirectoriesInVSCodeAuthenticatio
 import {
   COMMAND_AUTH0_AUTH,
   MEMENTO_RAZROO_ACCESS_TOKEN,
-  COMMAND_CANCEL_AUTH,
-  GENERATE_ANGULAR_COMPONENT,
-  GENERATE_ANGULAR_SERVICE,
-  COMMUNITY,
-  MEMENTO_RAZROO_ID_VS_CODE_TOKEN,
-  MEMENTO_RAZROO_USER_ID,
-  MEMENTO_RAZROO_ORG_ID,
-  GENERATE_ANGULAR_GUARD,
-  GENERATE_ANGULAR_TYPESCRIPT_INTERFACE,
-  GENERATE_ANGULAR_PIPE,
-  GENERATE_ANGULAR_DIRECTIVE,
+  COMMAND_CANCEL_AUTH
 } from './constants';
 import { createDisposableAuthServer } from './auth/local';
 import { Uri } from 'vscode';
@@ -29,7 +19,7 @@ import { getPackageJson, subscribeToGenerateVsCodeDownloadCodeSub } from './util
 import { EventEmitter } from 'stream';
 import { setWorkspaceState } from './utils/state.utils';
 import { getOrCreateAndUpdateIdToken } from './utils/token/token';
-import { createScaffold } from './utils/scaffold/scaffold';
+import { pushScaffoldCommands } from './utils/scaffold/push-scaffold-commands';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -67,36 +57,7 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.executeCommand('setContext', 'razroo-vscode-plugin:isAuthenticationInProgress', false);
   };
 
-  const generateAngularComponent = vscode.commands.registerCommand(
-    GENERATE_ANGULAR_COMPONENT,
-    async ({path}) => createScaffold('angular-15.0.0', 'angular-core', path, context, isProduction, 'component', packageJsonParams)
-  );
-  context.subscriptions.push(generateAngularComponent);
-  const generateAngularService = vscode.commands.registerCommand(
-    GENERATE_ANGULAR_SERVICE,
-    async ({path}) => createScaffold('angular-15.0.0', 'angular-core', path, context, isProduction, 'angular-service', packageJsonParams)
-  );
-  context.subscriptions.push(generateAngularService);
-  const generateAngularPipe = vscode.commands.registerCommand(
-    GENERATE_ANGULAR_PIPE,
-    async ({path}) => createScaffold('angular-15.0.0', 'angular-core', path, context, isProduction, 'angular-pipe', packageJsonParams)
-  );
-  context.subscriptions.push(generateAngularPipe);
-  const generateAngularGuard = vscode.commands.registerCommand(
-    GENERATE_ANGULAR_GUARD,
-    async ({path}) => createScaffold('angular-15.0.0', 'angular-core', path, context, isProduction, 'angular-guard', packageJsonParams)
-  );
-  context.subscriptions.push(generateAngularGuard);
-  const generateAngularDirective = vscode.commands.registerCommand(
-    GENERATE_ANGULAR_DIRECTIVE,
-    async ({path}) => createScaffold('angular-15.0.0', 'angular-core', path, context, isProduction, 'angular-directive', packageJsonParams)
-  );
-  context.subscriptions.push(generateAngularDirective);
-  const generateAngularTypescriptInterface = vscode.commands.registerCommand(
-    GENERATE_ANGULAR_TYPESCRIPT_INTERFACE,
-    async ({path}) => createScaffold('angular-15.0.0', 'angular-core', path, context, isProduction, 'typescript-interface', packageJsonParams)
-  );
-  context.subscriptions.push(generateAngularTypescriptInterface);
+  pushScaffoldCommands(context, isProduction, packageJsonParams);
 
   const auth0Authentication = vscode.commands.registerCommand(
     COMMAND_AUTH0_AUTH,
